requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.15.2
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
webdriver-manager==4.0.1
